// src/types/types.ts
export interface EdgeOtherMethods {
  readonly [name: string]: any
}

/** We frequently index things by pluginId, so provide a helper. */
export interface EdgePluginMap<Value> {
  [pluginName: string]: Value
}

type ArrayBufferView =
  | Int8Array
  | Int16Array
  | Int32Array
  | Uint8Array
  | Uint16Array
  | Uint32Array
  | Uint8ClampedArray
  | Float32Array
  | Float64Array
  | DataView

// ---------------------------------------------------------------------
// io types
// ---------------------------------------------------------------------

// Node.js randomBytes function:
export type EdgeRandomFunction = (bytes: number) => Uint8Array

// The scrypt function Edge expects:
export type EdgeScryptFunction = (
  data: Uint8Array,
  salt: Uint8Array,
  n: number,
  r: number,
  p: number,
  dklen: number
) => Promise<Uint8Array>

/**
 * The subset of the `fetch` options we guarantee to support.
 */
export interface EdgeFetchOptions {
  method?: string
  body?: ArrayBuffer | string
  headers?: { [header: string]: string }
}

/**
 * The subset of the `Headers` DOM object we guarantee to support.
 */
export interface EdgeFetchHeaders {
  forEach(
    callback: (value: string, name: string, self: EdgeFetchHeaders) => void,
    thisArg?: any
  ): void
  get(name: string): string | null
  has(name: string): boolean
}

/**
 * The subset of the `Response` DOM object we guarantee to support.
 */
export interface EdgeFetchResponse {
  readonly headers: EdgeFetchHeaders
  readonly ok: boolean
  readonly status: number
  arrayBuffer(): Promise<ArrayBuffer>
  json(): Promise<any>
  text(): Promise<string>
}

/**
 * The subset of the `fetch` DOM function we guarantee to support,
 * especially if we have to emulate `fetch` in weird environments.
 */
export type EdgeFetchFunction = (
  uri: string,
  opts?: EdgeFetchOptions
) => Promise<EdgeFetchResponse>

/**
 * The React Native Typescript types don't contain a WebSocket?!
 * This is copied from the spec WebIDL.
 */
export interface EdgeWebSocket {
  readonly url: string

  // ready state
  readonly CONNECTING: 0
  readonly OPEN: 1
  readonly CLOSING: 2
  readonly CLOSED: 3
  readonly readyState: string
  readonly bufferedAmount: string

  // networking
  onopen: Function
  onerror: Function
  onclose: Function
  readonly extensions: string
  readonly protocol: string
  close(code?: number, reason?: string): void

  // messaging
  onmessage: Function
  binaryType: 'arraybuffer'
  send(data: string | ArrayBuffer | ArrayBufferView): void
}

/**
 * Access to platform-specific resources.
 * The core never talks to the outside world on its own,
 * but always goes through this object.
 */
export interface EdgeIo {
  // Crypto:
  readonly random: EdgeRandomFunction
  readonly scrypt: EdgeScryptFunction

  // Local io:
  readonly disklet: Disklet
  readonly fetch: EdgeFetchFunction

  // This is only present if the platform has some way to avoid CORS:
  readonly fetchCors?: EdgeFetchFunction

  // Deprecated:
  // eslint-disable-next-line no-use-before-define
  readonly console: EdgeConsole
  readonly WebSocket: EdgeWebSocket
}
// ...
export interface EdgeNativeIo {
  [packageName: string]: EdgeOtherMethods
}
// ...
export interface EdgeConsole {
  error(...data: any[]): void
  info(...data: any[]): void
  warn(...data: any[]): void
}